+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## next
 
 - New: Read start and end systems from URL, e.g. domain.tdl/#Home;HED-GP;GE-8JV
+- Change: No more "Unknown Ansiblex", if an Ansiblex of a pair is missing, this connection is now ignored.
 - Small bug fixes
 
 ## 0.6.0